requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==3.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
iniconfig==1.1.1
packaging==21.3
pip==23.2.1
pluggy==1.0.0
psycopg2==2.8.6
py==1.11.0
pyparsing==3.0.4
pytest==7.1.2
pytz==2022.1
setuptools==68.0.0
sqlparse==0.4.4
tomli==1.2.2
wheel==0.38.4
